rename: ui 패키지 내 design-token 디렉토리명을 명확하게 design-token-stories 로 수정 (#38)
