+++ packages/ui/scripts/generate-tokens.ts
@@ -7,7 +7,10 @@ const __dirname = path.dirname(__filename);
 
 // CSS 파일 경로
 const TAILWIND_CONFIG_PATH = path.join(__dirname, '../../tailwind-config');
-const OUTPUT_PATH = path.join(__dirname, '../src/design-system/design-tokens/generated-tokens.ts');
+const OUTPUT_PATH = path.join(
+  __dirname,
+  '../src/design-system/design-tokens-stories/colors/generated-tokens.ts'
+);
 
 interface Variables {
   [key: string]: string;
