+++ packages/ui/src/design-system/design-token-stories/colors/generated-tokens.ts
