+++ packages/ui/src/utils/storybook/color-token-utils.ts
@@ -1,7 +1,7 @@
 import {
   primitiveColors,
   semanticColorCSSValues,
-} from '../../design-system/design-tokens/generated-tokens';
+} from '../../design-system/design-token-stories/colors/generated-tokens';
 
 // 색상 타입별 매핑
 const COLOR_TYPE_MAPPING = {
